requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.1
aiohttp-retry==2.9.1
aiosignal==1.4.0
annotated-types==0.7.0
anyio==4.11.0
attrs==25.4.0
Authlib==1.6.5
bandit==1.8.6
beautifulsoup4==4.14.2
certifi==2025.10.5
cffi==2.0.0
cfgv==3.4.0
charset-normalizer==3.4.4
click==8.3.0
coverage==7.11.0
cryptography==46.0.3
distlib==0.4.0
dparse==0.6.4
filelock==3.20.0
firecrawl-py>=4.5.0
Flask==3.0.0
flask-cors==4.0.2
frozenlist==1.8.0
greenlet==3.2.4
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
identify==2.6.15
idna==3.11
iniconfig==2.3.0
Jinja2==3.1.6
joblib==1.5.2
lxml==6.0.2
markdown-it-py==4.0.0
MarkupSafe==3.0.3
marshmallow==4.0.1
mdurl==0.1.2
multidict==6.7.0
mypy==1.18.2
mypy_extensions==1.1.0
nltk==3.9.2
nodeenv==1.9.1
packaging==25.0
pathspec==0.12.1
platformdirs==4.5.0
playwright==1.55.0
pluggy==1.6.0
pre_commit==4.3.0
propcache==0.4.1
psutil==7.1.1
pycparser==2.23
pydantic==2.12.3
pydantic_core==2.41.4
pyee==13.0.0
Pygments==2.19.2
PyJWT==2.10.1
pytest==8.4.2
pytest-cov==7.0.0
pytest-html==4.1.1
pytest-metadata==3.1.1
pytest-mock==3.15.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
PyYAML==6.0.3
rapidfuzz>=3.10.0
regex==2025.10.23
requests==2.32.5
rich==14.2.0
ruamel.yaml==0.18.16
ruamel.yaml.clib==0.2.14
ruff==0.14.1
safety==3.6.2
safety-schemas==0.0.16
setuptools==80.9.0
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
soupsieve==2.8
SQLAlchemy==2.0.44
stevedore==5.5.0
tenacity==9.1.2
tomlkit==0.13.3
tqdm==4.67.1
twilio==9.8.4
typer==0.20.0
types-beautifulsoup4==4.12.0.20250516
types-html5lib==1.1.11.20251014
types-requests==2.32.4.20250913
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.5.0
virtualenv==20.35.3
yarl==1.22.0
